+++ src/screens/addProfile/AddProfile.tsx
@@ -5,7 +5,7 @@ import { ImageURISource } from 'react-native'
 import { useMutation } from 'react-query'
 import { useRecoilValue, useSetRecoilState } from 'recoil'
 
-import { checkNicknameValid } from '@/apis/user'
+import { checkNicknameValid } from '@/apis/user/User'
 import XMark from '@/assets/icons/x_mark.svg'
 import { BoxButton } from '@/components/buttons/boxButton/BoxButton'
 import DefaultContainer from '@/components/containers/defaultContainer/DefaultContainer'
