+++ src/screens/contentDetail/ContentDetail.tsx
@@ -5,8 +5,8 @@ import { RouteProp, useNavigation } from '@react-navigation/native'
 import { AxiosError } from 'axios'
 import { useMutation, useQuery, useQueryClient } from 'react-query'
 
-import { postBlock } from '@/apis/block'
-import { deleteContents, getContents } from '@/apis/content'
+import { postBlock } from '@/apis/block/Block'
+import { deleteContents, getContents } from '@/apis/content/Content'
 import { defaultImages } from '@/assets'
 import DefaultContainer from '@/components/containers/defaultContainer/DefaultContainer'
 import DefaultScrollContainer from '@/components/containers/defaultScrollContainer/DefaultScrollContainer'
