+++ src/screens/main/home/Home.tsx
@@ -7,7 +7,7 @@ import { useInfiniteQuery, useQuery, useQueryClient } from 'react-query'
 import { useRecoilState, useRecoilValue } from 'recoil'
 
 import { getHomeArchivingList } from '@/apis/archiving/ArchivingList'
-import { getUser } from '@/apis/user'
+import { getUser } from '@/apis/user/User'
 import { defaultImages } from '@/assets'
 import SearchButton from '@/components/buttons/searchButton/SearchButton'
 import { ArchivingCard } from '@/components/cards/archivingCard/ArchivingCard'
