+++ src/screens/myAccount/MyAccount.tsx
@@ -5,8 +5,8 @@ import { useNavigation } from '@react-navigation/native'
 import { ImageSourcePropType, ImageURISource, TouchableOpacity, View } from 'react-native'
 import { useMutation, useQuery, useQueryClient } from 'react-query'
 
-import { deleteWithdrawal } from '@/apis/auth'
-import { getUserInfo, postUserInfo } from '@/apis/user'
+import { deleteWithdrawal } from '@/apis/auth/Auth'
+import { getUserInfo, postUserInfo } from '@/apis/user/User'
 import { defaultImages } from '@/assets'
 import PencilIcon from '@/assets/icons/pencil.svg'
 import DefaultContainer from '@/components/containers/defaultContainer/DefaultContainer'
