+++ src/screens/mypage/Mypage.tsx
@@ -7,8 +7,8 @@ import LinearGradient from 'react-native-linear-gradient'
 import { Shadow } from 'react-native-shadow-2'
 import { useMutation, useQuery, useQueryClient } from 'react-query'
 
-import { logout } from '@/apis/auth'
-import { getUser } from '@/apis/user'
+import { logout } from '@/apis/auth/Auth'
+import { getUser } from '@/apis/user/User'
 import { defaultImages } from '@/assets'
 import LeftArrowIcon from '@/assets/icons/left_arrow.svg'
 import DefaultContainer from '@/components/containers/defaultContainer/DefaultContainer'
